+++ source/core/screen.hpp
@@ -1,11 +1,11 @@
 #pragma once
 
 struct AspectRatio {
-	float width;
-	float height;
+	int width;
+	int height;
 
 	AspectRatio();
-	AspectRatio(float _width, float _height);
+	AspectRatio(int _width, int _height);
 
 	float aspect() const;
 
@@ -21,6 +21,6 @@ struct Screen {
 	void set(float _width, float _height);
 	void set(float _height, const AspectRatio & aspect_ratio);
 	void set_coordinates();
-	void adjust(float lines);
+	void adjust(int lines);
 
 };
